requirements.txt
# PyTorch - Instala a versão com CUDA por padrão (se disponível)
# Para instalar apenas CPU: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
jupyter>=1.0.0
scikit-learn>=1.0.0
